restructured the hierarchy and cleaned up the code.Code is now ready to be shipped to qa
