+++ Stylist/Stylist/Controllers/Client Files/ClientBookingsController.swift
@@ -1,0 +1,20 @@
+//
+//  ClientBookingsController.swift
+//  Stylist
+//
+//  Created by Ashli Rankin on 4/5/19.
+//  Copyright © 2019 Ashli Rankin. All rights reserved.
+//
+
+import UIKit
+
+class ClientBookingsController: UITableViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+      
+    }
+
+
+}
