+++ Stylist/Stylist/Controllers/Client Files/ClientSearchController.swift
@@ -1,0 +1,20 @@
+//
+//  ClientSearchController.swift
+//  Stylist
+//
+//  Created by Ashli Rankin on 4/5/19.
+//  Copyright © 2019 Ashli Rankin. All rights reserved.
+//
+
+import UIKit
+
+class ClientSearchController: UITableViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+     
+    }
+
+  
+}
