+++ Stylist/Stylist/Controllers/Entrance Controllers/CreateViewController.swift
@@ -12,15 +12,14 @@ class CreateViewController: BaseViewController {
 
   @IBOutlet weak var emailTextfield: UITextField!
   @IBOutlet weak var passwordTextfield: UITextField!
-  @IBOutlet weak var accountTypeControl: UISegmentedControl!
-  
- 
+
   let authService = AuthService()
   
   override func viewDidLoad() {
         super.viewDidLoad()
     authService.authserviceCreateNewAccountDelegate = self
-
+      emailTextfield.delegate = self
+      passwordTextfield.delegate = self
     }
     
   @IBAction func createProfilePressed(_ sender: UIButton) {
@@ -30,14 +29,9 @@ class CreateViewController: BaseViewController {
       return
     }
    authService.createNewAccount(email: email, password: password)
-  
-   
+
   }
-  
-  @IBAction func segmentedControlPressed(_ sender: UISegmentedControl) {
-    
-  }
-  
+ 
   private func presentOnboardingScreen(){
      let onbordingScreen = UIStoryboard(name: "Entrance", bundle: nil).instantiateViewController(withIdentifier: "OnboardingTableViewController")
     let navigationController = UINavigationController(rootViewController: onbordingScreen)
@@ -48,9 +42,12 @@ class CreateViewController: BaseViewController {
     navigationController.navigationBar.shadowImage = UIImage()
     self.present(navigationController, animated: true, completion: nil)
   }
+  
+ 
 }
 extension CreateViewController:UITextFieldDelegate{
   func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+    textField.resignFirstResponder()
     return true
     
   }
