+++ Stylist/Stylist/Controllers/Service Provider Files/ProviderDetailController.swift
@@ -9,7 +9,9 @@
 import UIKit
 
 class ProviderDetailController: UITableViewController {
+  
     @IBOutlet weak var scrollView: UIScrollView!
+  
     lazy var providerDetailHeader = UserDetailView(frame: CGRect(x: 0, y: 0, width: view.bounds.width, height: 300))
     @IBOutlet weak var collectionView: UICollectionView!
     var buttons = ["Bio", "Portfolio","Availability", "Reviews"] {
