+++ Stylist/Stylist/Controllers/Service Provider Files/ServiceDetailViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ServiceDetailViewController.swift
+//  Stylist
+//
+//  Created by Ashli Rankin on 4/5/19.
+//  Copyright © 2019 Ashli Rankin. All rights reserved.
+//
+
+import UIKit
+
+class ServiceDetailViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+
+}
