+++ Stylist/Stylist/Controllers/Service Provider Files/ServiceMessageViewController.swift
@@ -1,0 +1,21 @@
+//
+//  ServiceMessageViewController.swift
+//  Stylist
+//
+//  Created by Ashli Rankin on 4/5/19.
+//  Copyright © 2019 Ashli Rankin. All rights reserved.
+//
+
+import UIKit
+
+class ServiceMessageViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+ 
+
+}
