+++ Stylist/Stylist/Controllers/Service Provider Files/ServiceUpcomingTableViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ServiceUpcomingTableViewController.swift
+//  Stylist
+//
+//  Created by Ashli Rankin on 4/5/19.
+//  Copyright © 2019 Ashli Rankin. All rights reserved.
+//
+
+import UIKit
+
+class ServiceUpcomingTableViewController: UITableViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+  }
+
+
+}
