--- Stylist/Stylist/Controllers/ServiceDetailViewController.swift
@@ -1,30 +1,0 @@
-//
-//  ServiceDetailViewController.swift
-//  Stylist
-//
-//  Created by Ashli Rankin on 4/5/19.
-//  Copyright © 2019 Ashli Rankin. All rights reserved.
-//
-
-import UIKit
-
-class ServiceDetailViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
