+++ Stylist/Stylist/SupportingFiles/Extensions/UIViewController+Extensions.swift
@@ -9,8 +9,6 @@
 import Foundation
 import UIKit
 
-
-
 extension UIViewController {
     public func setupScrollviewController(scrollView: UIScrollView, views: [UIView]) {
         scrollView.isPagingEnabled = true
