+++ Stylist/Stylist/SupportingFiles/Firebase Services/AuthService.swift
