+++ Stylist/Stylist/Views/Cells/ProviderDetailCustomCell.swift
